Modularized Sidebar call, home page, and main button component
