+++ components/basecomponents/MainCard.jsx
@@ -1,0 +1,6 @@
+export default function MainCard ({ children, ...props }) {
+  return (
+    <div></div>
+
+  )
+}
