+++ pages/[username]/calendar.js
@@ -1,1 +1,7 @@
 // Calendar Widget
+
+export default function Calendar() {
+  return (
+    <div className='pageData'>My Calendar</div>
+  )
+}
