+++ pages/[username]/class-catalog.js
@@ -1,1 +1,10 @@
 // ClassCatalog widget
+import MainButton from '../../components/basecomponents/MainButton.jsx';
+
+export default function ClassCatalog() {
+  return (
+    <div className='pageData'>Class Catalog
+    <MainButton value="Search"/>
+    </div>
+  )
+}
