+++ pages/[username]/index.js
@@ -1,3 +1,7 @@
 // Dashboard
 
-export default function Dashboard() {}
+export default function Dashboard() {
+  return (
+    <div className='pageData'>Dashboard</div>
+  )
+}
