+++ pages/[username]/my-profile.js
@@ -1,3 +1,8 @@
 // Profile View
 
-export default function MyProfile(props) {}
+export default function MyProfile() {
+  return (
+    <div className='pageData'>My Profile</div>
+  )
+}
+
