+++ pages/[username]/notifications.js
@@ -1,1 +1,7 @@
 // Notifications Page
+
+export default function Notifications() {
+  return (
+    <div className='pageData'>Notifications</div>
+  )
+}
