+++ utils/constants/index.js
@@ -2,5 +2,4 @@
 
 const defaultProfilePic = "https://www.vhv.rs/dpng/d/426-4264903_user-avatar-png-picture-avatar-profile-dummy-transparent.png"
 
-
 export default defaultProfilePic;
