+++ utils/styles/NavLayoutStyles/HomePage.module.css
@@ -48,7 +48,7 @@
 
 .reporter {
   width: 100vw;
-text-align: center;
+  text-align: center;
   margin-top: 30vh;
   font-size: 25px;
   font-weight: 100;
